# tracing-context: optimize Visit impl