--- a/metrics-tracing-context/.gitignore
+++ b/metrics-tracing-context/.gitignore
@@ -1,0 +1,3 @@
+/target
+**/*.rs.bk
+Cargo.lock