--- a/metrics-tracing-context/Cargo.toml
+++ b/metrics-tracing-context/Cargo.toml
@@ -15,12 +15,21 @@ readme = "README.md"
 categories = ["development-tools::debugging"]
 keywords = ["metrics", "tracing"]
 
+[lib]
+bench = false
+
+[[bench]]
+name = "visit"
+harness = false
+
 [dependencies]
 metrics = { version = "0.13.0-alpha.1", path = "../metrics", features = ["std"] }
 metrics-util = { version = "0.4.0-alpha.1", path = "../metrics-util" }
 tracing = "0.1"
 tracing-core = "0.1"
 tracing-subscriber = "0.2"
+itoa = "0.4"
 
 [dev-dependencies]
+criterion = "0.3"
 parking_lot = "0.11"