--- a/metrics-tracing-context/src/tracing_integration.rs
+++ b/metrics-tracing-context/src/tracing_integration.rs
@@ -6,11 +6,37 @@ use tracing_core::span::{Attributes, Id, Record};
 use tracing_core::{field::Visit, Dispatch, Field, Subscriber};
 use tracing_subscriber::{layer::Context, registry::LookupSpan, Layer};
 
-struct Labels(Vec<Label>);
+/// Per-span extension for collecting labels from fields.
+///
+/// Hidden from documentation as there is no need for end users to ever touch this type, but it must
+/// be public in order to be pulled in by external benchmark code.
+#[doc(hidden)]
+pub struct Labels(pub Vec<Label>);
 
 impl Visit for Labels {
     fn record_str(&mut self, field: &Field, value: &str) {
         let label = Label::new(field.name(), value.to_owned());
+        self.0.push(label);
+    }
+
+    fn record_bool(&mut self, field: &Field, value: bool) {
+        let label = Label::new(field.name(), if value { "true" } else { "false " });
+        self.0.push(label);
+    }
+
+    fn record_i64(&mut self, field: &Field, value: i64) {
+        // Maximum length is 20 characters but 32 is a nice power-of-two number.
+        let mut s = String::with_capacity(32);
+        itoa::fmt(&mut s, value).expect("failed to format/write i64");
+        let label = Label::new(field.name(), s);
+        self.0.push(label);
+    }
+
+    fn record_u64(&mut self, field: &Field, value: u64) {
+        // Maximum length is 20 characters but 32 is a nice power-of-two number.
+        let mut s = String::with_capacity(32);
+        itoa::fmt(&mut s, value).expect("failed to format/write u64");
+        let label = Label::new(field.name(), s);
         self.0.push(label);
     }