--- a/metrics-util/benches/bucket.rs
+++ b/metrics-util/benches/bucket.rs
@@ -1,10 +1,5 @@
-#[macro_use]
-extern crate criterion;
-
-#[macro_use]
-extern crate lazy_static;
-
-use criterion::{Benchmark, Criterion, Throughput};
+use criterion::{criterion_group, criterion_main, Benchmark, Criterion, Throughput};
+use lazy_static::lazy_static;
 use metrics_util::AtomicBucket;
 
 lazy_static! {