When it was originally written, `metrics-tracing-context` used only a `record_str` and `record_debug` override for the `tracing::field::Visit` implementation, which allowed full functionality but wasn't entirely optimized.  This PR simply fills in the other `record_*` functions (except for `record_error`) with optimized implementations.

Performance increases:
- boolean fields: ~75% faster (54ns -> 15ns)
- integer fields: ~40% faster (70ns -> 40ns)